# api.yaml
openapi: 3.0.0
info:
  title: Video Game Bulletin (VGB) API
  description: RESTful API for managing video game releases, user interactions, and reviews
  version: 1.0.0
  contact:
    name: Project VAUYL
    email: vgb@example.com

servers:
  - url: http://localhost:3000/api
    description: Development server
  - url: https://video-game-bulletin.firebaseapp.com/api
    description: Production server

tags:
  - name: Games
    description: Game release information endpoints
  - name: Users
    description: User authentication and management
  - name: Comments
    description: User comments on games
  - name: Favorites
    description: User favorite games management

paths:
  /releases:
    get:
      tags:
        - Games
      summary: Get all upcoming and recent game releases
      description: Retrieve list of games for calendar display
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Game'
        '500':
          description: Server error

  /games:
    post:
      tags:
        - Games
      summary: Add new game (Admin only)
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GameInput'
      responses:
        '201':
          description: Game created successfully
        '401':
          description: Unauthorized
        '403':
          description: Forbidden - Admin only

  /games/{id}:
    get:
      tags:
        - Games
      summary: Get specific game details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Game details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Game'
        '404':
          description: Game not found

    put:
      tags:
        - Games
      summary: Update game information (Admin only)
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GameInput'
      responses:
        '200':
          description: Game updated successfully
        '401':
          description: Unauthorized
        '403':
          description: Forbidden - Admin only
        '404':
          description: Game not found

    delete:
      tags:
        - Games
      summary: Delete game (Admin only)
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Game deleted successfully
        '401':
          description: Unauthorized
        '403':
          description: Forbidden - Admin only
        '404':
          description: Game not found

  /users:
    post:
      tags:
        - Users
      summary: Register new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRegistration'
      responses:
        '201':
          description: User created successfully
        '400':
          description: Invalid input

  /users/login:
    post:
      tags:
        - Users
      summary: User login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                  user:
                    $ref: '#/components/schemas/User'
        '401':
          description: Invalid credentials

  /users/{userId}/favorites:
    get:
      tags:
        - Favorites
      summary: Get user's favorite games
      security:
        - BearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: List of favorite games
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Game'

    post:
      tags:
        - Favorites
      summary: Add game to favorites
      security:
        - BearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                game_id:
                  type: string
      responses:
        '201':
          description: Game added to favorites
        '401':
          description: Unauthorized

  /games/{id}/comments:
    get:
      tags:
        - Comments
      summary: Get comments for a game
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: List of comments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Comment'

    post:
      tags:
        - Comments
      summary: Post comment on game
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                content:
                  type: string
      responses:
        '201':
          description: Comment posted successfully
        '401':
          description: Unauthorized

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    Game:
      type: object
      properties:
        game_id:
          type: string
        title:
          type: string
        release_date:
          type: string
          format: date
        platforms:
          type: array
          items:
            type: string
        specifications:
          type: object
        genre:
          type: string
        description:
          type: string
        created_at:
          type: string
          format: date-time

    GameInput:
      type: object
      required:
        - title
        - release_date
        - platforms
        - genre
      properties:
        title:
          type: string
        release_date:
          type: string
          format: date
        platforms:
          type: array
          items:
            type: string
        specifications:
          type: object
        genre:
          type: string
        description:
          type: string

    User:
      type: object
      properties:
        user_id:
          type: string
        username:
          type: string
        email:
          type: string
          format: email
        is_admin:
          type: boolean
        created_at:
          type: string
          format: date-time

    UserRegistration:
      type: object
      required:
        - username
        - email
        - password
      properties:
        username:
          type: string
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6

    Comment:
      type: object
      properties:
        comment_id:
          type: string
        game_id:
          type: string
        user_id:
          type: string
        username:
          type: string
        content:
          type: string
        timestamp:
          type: string
          format: date-time